requirements.txt
dask~=2023.6.0
numpy~=1.23.5
zarr~=2.14.2
aicsimageio~=4.11.0
tifffile~=2023.4.12
xmltodict~=0.13.0
rich~=13.4.2
opencv-python~=4.5.5.62
shapely~=2.0.1
Pillow~=9.5.0
matplotlib~=3.8.0
geojson~=3.0.1
napari~=0.4.17
scikit-image~=0.19.3
pydantic~=1.10.9
numcodecs~=0.11.0
imagecodecs~=2023.3.16
scikit-learn~=1.3.0
pandas~=1.5.3
scipy~=1.10.1
threadpoolctl~=3.2.0
tqdm~=4.65.0
setuptools~=65.5.1